Fixed the bug where data points were not getting removed when the data changed. This was due to enum type dependency properties not being supported in animations in WinRT.

+++ WinRTXamlToolkit.Controls.DataVisualization/Charting/DataPoint/BubbleDataPoint.cs
@@ -65,7 +65,7 @@ namespace WinRTXamlToolkit.Controls.DataVisualization.Charting
                 handler(this, new RoutedPropertyChangedEventArgs<double>(oldValue, newValue));
             }
 
-            if (this.State == DataPointState.Created)
+            if (this.State == (int)DataPointState.Created)
             {
                 this.ActualSize = newValue;
             }
