+++ WinRTXamlToolkit.Controls.DataVisualization/Charting/DataPoint/PieDataPoint.cs
@@ -299,7 +299,7 @@ namespace WinRTXamlToolkit.Controls.DataVisualization.Charting
                 {
                     handler(this, new RoutedPropertyChangedEventArgs<double>(oldValue, newValue));
                 }
-                if (this.State == DataPointState.Created)
+                if (this.State == (int)DataPointState.Created)
                 {
                     ActualOffsetRatio = newValue;
                 }
@@ -367,7 +367,7 @@ namespace WinRTXamlToolkit.Controls.DataVisualization.Charting
                     handler(this, new RoutedPropertyChangedEventArgs<double>(oldValue, newValue));
                 }
 
-                if (this.State == DataPointState.Created)
+                if (this.State == (int)DataPointState.Created)
                 {
                     ActualRatio = newValue;
                 }
